Move shadow to its own view
